+++ examples/branch-protection/main.tf
@@ -34,7 +34,7 @@ locals {
 module "branch_protection" {
   source                 = "./../../"
   name                   = "example-branch-protection"
-  description            = "A Terraform repository example"
+  description            = "A github repository example created using terraform"
   license_template       = "apache-2.0"
   allow_squash_merge     = true
   branch                 = ["dev", "pre", "prd"]
@@ -44,6 +44,10 @@ module "branch_protection" {
   branch_protection_version = {
     use_branch_protection    = true
     use_branch_protection_v3 = false
+  }
+  required_status_checks = {
+    strict   = true
+    contexts = ["checkov-scan / checkov-scan"]
   }
   template = {
     owner      = "boldlink"