+++ examples/complete/main.tf
@@ -27,8 +27,8 @@ resource "github_team" "maintain" {
 
 module "complete" {
   source                 = "./../../"
-  name                   = "example-complete"
-  description            = "A Terraform repository example."
+  name                   = "example-complete-repository"
+  description            = "A github repository example created using terraform."
   license_template       = "apache-2.0"
   allow_squash_merge     = true
   branch                 = ["dev", "pre", "prd"]
@@ -37,6 +37,19 @@ module "complete" {
   require_signed_commits = true
   homepage_url           = "https://boldlink.io"
   pattern                = "develop"
+  visibility             = "private"
+  has_issues             = true
+  has_projects           = true
+  has_wiki               = true
+  is_template            = false
+  allow_merge_commit     = false
+  delete_branch_on_merge = true
+  has_downloads          = false
+  auto_init              = true
+  archive_on_destroy     = false
+  vulnerability_alerts   = true
+  archived               = false
+  topics                 = []
   template = {
     owner      = "boldlink"
     repository = "terraform-module-template"
@@ -47,16 +60,19 @@ module "complete" {
   }
 
   branch_protection_version = {
-    use_branch_protection    = true
-    use_branch_protection_v3 = false
+    use_branch_protection    = false
+    use_branch_protection_v3 = true
   }
-  required_pull_request_reviews = {
+  enforce_admins                  = true
+  require_conversation_resolution = true
+
+  required_pull_request_reviews_v3 = {
     dismiss_stale_reviews           = true
     require_code_owner_reviews      = true
     required_approving_review_count = 2
     restrict_dismissals             = true
-    dismissal_restrictions          = []
-    pull_request_bypassers          = [github_team.admin.node_id]
+    dismissal_teams                 = [github_team.admin.node_id]
+    dismissal_users                 = ["ndegwajohn"]
   }
   restrictions = {
     users = []
@@ -67,14 +83,16 @@ module "complete" {
     apps = []
   }
 
-  required_status_checks = {
-    strict   = true
-    contexts = ["checkov-scan / checkov-scan"]
+  required_status_checks_v3 = {
+    strict = true
+    checks = [
+      "ci/check"
+    ]
   }
   issue_label = {
     patch = {
       color       = "ff0000"
-      description = "Sample label for complete example repo"
+      description = "Example label for complete repository"
     }
   }