+++ examples/minimum/main.tf
@@ -1,5 +1,5 @@
 module "minimum" {
   source      = "./../../"
   name        = "example-minimum"
-  description = "Terraform sample github repository "
+  description = "Example repository created using terraform."
 }