+++ variables.tf
@@ -158,11 +158,14 @@ variable "teams" {
 }
 
 variable "required_status_checks" {
-  type = object({
-    strict   = bool
-    contexts = list(string)
-  })
+  type        = any
   description = "Enforce restrictions for required status checks, `strict` (Optional) Require branches to be up to date before merging. Defaults to false `contexts` (Optional) The list of status checks to require in order to merge into this branch. No status checks are required by default."
+  default     = null
+}
+
+variable "required_status_checks_v3" {
+  type        = any
+  description = "Enforce restrictions for required status checks, `strict` (Optional) Require branches to be up to date before merging. Defaults to false `checks` (Optional) The list of status checks to require in order to merge into this branch. No status checks are required by default."
   default     = null
 }