+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: CKV_GIT_5: "Ensure at least two approving reviews for PRs"
 - feat: bypass pull request allowances in branch protection v3
 - show repository pages example
+- feat: add checks in required status checks
+- fix: pull_request_bypassers in branch protection
 
 
 ## [1.3.1] - 2023-11-24
@@ -18,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - showcased the following attributes in complete example visibility, has_issues,has_projects,has_wiki, is_template, allow_merge_commit, delete_branch_on_merge, has_downloads, auto_init, archive_on_destroy, vulnerability_alerts, archived
 - added branch protection V3 in complete example
 - added a depends on to github labels resource
+- fix: required_status_checks in branch protection v3
 
 ## [1.3.0] - 2022-11-21
 ### Changes